movies_admin/requirements.txt
django==4.2.5
django-debug-toolbar==4.3.0
django-split-settings==1.2.0
flake8==5.0.4
flake8-blind-except==0.2.1
flake8-broken-line==0.6.0
flake8-bugbear==22.9.23
flake8-builtins==2.0.0
flake8-class-attributes-order==0.1.3
flake8-cognitive-complexity==0.1.0
flake8-commas==2.1.0
flake8-comprehensions==3.10.0
flake8-debugger==4.1.2
flake8-eradicate==1.4.0
flake8-functions==0.0.7
flake8-isort==5.0.0
flake8-mutable==1.2.0
flake8-print==5.0.0
flake8-pytest==1.4
flake8-pytest-style==1.6.0
flake8-quotes==3.3.1
flake8-string-format==0.3.0
flake8-variables-names==0.0.5
psycopg2==2.9.9
python-dotenv==1.0.1